Refactor: Move `model` package to `internal/model`
Relocated the `model` package to `internal/model` for better module encapsulation and adherence to Go project structure. Adjusted all relevant imports and usages across the codebase to reflect this change. Additionally, extracted `MetaData` into a new `metadata` package for improved modularity.
