+++ internal/config/esconfig/postgres_repository_test.go
@@ -3,8 +3,8 @@ package esconfig
 import (
 	"fmt"
 	"github.com/jmoiron/sqlx"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/tests"
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
 	"testing"
 )
