+++ internal/config/esconfig/repository.go
@@ -3,7 +3,7 @@ package esconfig
 import (
 	"sync"
 
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 )
 
 // Repository is a storage interface which can be implemented by multiple backend
