+++ internal/explainer/action/postgres_repository_test.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/jmoiron/sqlx"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/explainer/rootcause"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/rule"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/tests"
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
 )
 
 func dbInit(dbClient *sqlx.DB, t *testing.T) {
