+++ internal/explainer/feedback.go
@@ -11,7 +11,7 @@ import (
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/explainer/action"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/explainer/issues"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/explainer/rootcause"
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 	"github.com/myrteametrics/myrtea-sdk/v5/postgres"
 )
