+++ internal/explainer/rootcause/postgres_repository.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/jmoiron/sqlx"
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 )
 
 // PostgresRepository is a repository containing the RootCause definition based on a PSQL database and
