+++ internal/handler/actions_handlers_test.go
@@ -13,9 +13,9 @@ import (
 	"github.com/jmoiron/sqlx"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/explainer/action"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/explainer/rootcause"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/rule"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/tests"
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
 )
 
 func actionsDBInit(dbClient *sqlx.DB, t *testing.T) {
