+++ internal/handler/rootcauses_handlers_test.go
@@ -12,9 +12,9 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/jmoiron/sqlx"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/explainer/rootcause"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/rule"
 	"github.com/myrteametrics/myrtea-engine-api/v5/internal/tests"
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
 )
 
 func rootCausesDBInit(dbClient *sqlx.DB, t *testing.T) {
