+++ internal/handler/utils.go
@@ -21,7 +21,7 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 	"go.uber.org/zap"
 )
