+++ internal/model/cases.go
