+++ internal/model/close.go
