+++ internal/model/connectorconfig.go
