+++ internal/model/evaluator.go
