+++ internal/model/facts.go
