+++ internal/model/search.go
