+++ internal/model/task.go
@@ -31,15 +31,6 @@ type InputTask struct {
 	Rule               RuleInput
 }
 
-// MetaData struct to represent a situation metadata
-type MetaData struct {
-	Key         string      `json:"key"`
-	Value       interface{} `json:"value"`
-	RuleID      int64       `json:"ruleId"`
-	RuleVersion int64       `json:"ruleVersion"`
-	CaseName    string      `json:"caseName"`
-}
-
 // struct Key for an map MetaData
 type Key struct {
 	SituationID         int64
