+++ internal/utils/queryutils/utils_test.go
@@ -3,7 +3,7 @@ package queryutils
 import (
 	"testing"
 
-	"github.com/myrteametrics/myrtea-engine-api/v5/pkg/model"
+	"github.com/myrteametrics/myrtea-engine-api/v5/internal/model"
 )
 
 func TestAppendSearchOptions(t *testing.T) {
